DetailPageのフォルダを分けた
