+++ lib/pages/DetailPage/detail_page.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import '../../models/metronome_model.dart';
+import 'detail_bottom_bar.dart';
+
+class DetailPage extends StatelessWidget {
+  final int bpm;
+  final String title;
+
+  DetailPage({Key key, this.bpm, this.title}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Consumer<MetronomeModel>(builder: (_, model, __) {
+      return Scaffold(
+        appBar: AppBar(
+          centerTitle: true,
+          leading: IconButton(
+              icon: Icon(Icons.arrow_back_ios),
+              onPressed: () {
+                Navigator.of(context).pop();
+                model.forceStop();
+              }),
+          title: Text(title),
+          actions: <Widget>[],
+        ),
+        body: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Text(
+                'テキスト',
+              ),
+              Text(bpm.toString()),
+            ],
+          ),
+        ),
+        bottomNavigationBar: detailBottomBar(context, model),
+      );
+    });
+  }
+}
